source/requirement.txt
Django==1.11.23
django-cors-headers==3.2.0
djangorestframework==3.10.3
pytz==2019.3
